Added Proposal domain type with signing - tests fail for now
