+++ tendermint/src/proposal/msg_type.rs
@@ -1,0 +1,45 @@
+use tendermint_proto::DomainType;
+use std::convert::TryFrom;
+use crate::{Error, Kind};
+use serde::{Serialize, Serializer, Deserialize, Deserializer};
+use serde::de::Error as _;
+
+/// Types of proposals
+#[repr(u8)]
+#[derive(Copy, Clone, Debug, Eq, Hash, PartialEq)]
+pub enum Type {
+    /// Regular Proposal
+    Proposal = 32,
+}
+
+impl DomainType<i32> for Type {}
+
+impl TryFrom<i32> for Type {
+    type Error = Error;
+
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
+        match value {
+            32 => Ok(Type::Proposal),
+            _ => Err(Kind::InvalidMessageType.into()),
+        }
+    }
+}
+
+impl From<Type> for i32 {
+    fn from(value: Type) -> Self {
+        value as i32
+    }
+}
+
+impl Serialize for Type {
+    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+        i32::from(*self).serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for Type {
+    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+        let byte = i32::deserialize(deserializer)?;
+        Type::try_from(byte).map_err(|_| D::Error::custom(format!("invalid proposal type: {}", byte)))
+    }
+}
